Module neptune::circuit2

Structs

PoseidonCircuit2

Circuit for Poseidon hash.

Enums

Elt

Similar to num::Num, we use Elt to accumulate both values and linear combinations, then eventually extract into a num::AllocatedNum, enforcing that the linear combination corresponds to the result.

Functions

mul_pre_sum

Calculates a * (b + to_add) — and enforces that constraint.

mul_sum

Calculates (a * (pre_add + b)) + post_add — and enforces that constraint.

poseidon_hash_allocated

Create circuit for Poseidon hash, returning an allocated Num at the cost of one constraint.

poseidon_hash_num

Create circuit for Poseidon hash, minimizing constraints by returning an unallocated Num.

square_sum

Calculates square of sum and enforces that constraint.